Correcciones

+++ backend_broasteria/src/productos/productos.service.ts
@@ -1,4 +1,4 @@
-import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
+import {ConflictException,Injectable,NotFoundException,} from '@nestjs/common';
 import { CreateProductoDto } from './dto/create-producto.dto';
 import { UpdateProductoDto } from './dto/update-producto.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -8,31 +8,32 @@ import { Categoria } from 'src/categorias/entities/categoria.entity';
 
 @Injectable()
 export class ProductosService {
-
-  constructor(@InjectRepository(Producto) private productosRepository: Repository<Producto>) { }
-
-
+  constructor(
+    @InjectRepository(Producto) private productosRepository: Repository<Producto>,
+  ) {}
 
   async create(createProductoDto: CreateProductoDto): Promise<Producto> {
-    const existe = await this.productosRepository.findOneBy({
-      nombre: createProductoDto.nombre.trim(),
-      categoria: { id: createProductoDto.idCategoria },
+    const existe = await this.productosRepository.findOne({
+      where: {
+       nombre: createProductoDto.nombre,
+        categoria: { id: createProductoDto.idCategoria },
+      },
+      relations: ['categoria'],
     });
     if (existe) {
-      throw new ConflictException('El producto ya existe');
+      throw new ConflictException('La producto ya existe');
     }
-    return this.productosRepository.save({
+    const producto = this.productosRepository.create({
       nombre: createProductoDto.nombre.trim(),
       descripcion: createProductoDto.descripcion.trim(),
       precioUnitario: createProductoDto.precioUnitario,
       stock: createProductoDto.stock,
-      categoria: { id: createProductoDto.idCategoria },
-
+      categoria: { id: createProductoDto.idCategoria } as Categoria,
     });
+    return this.productosRepository.save(producto);
   }
 
   async findAll(): Promise<Producto[]> {
-
     return this.productosRepository.find({ relations: ['categoria'] });
   }
 
@@ -42,20 +43,26 @@ export class ProductosService {
       relations: ['categoria'],
     });
     if (!producto) {
-      throw new NotFoundException(`El producto con el id  ${id} no existe`);
+      throw new NotFoundException(`La producto con el id ${id} no existe`);
     }
     return producto;
   }
 
   async update(id: number, updateProductoDto: UpdateProductoDto): Promise<Producto> {
     const producto = await this.findOne(id);
+    if (!producto) {
+      throw new NotFoundException(`La producto con el id ${id} no existe`);
+    }
     const actualizarProducto = Object.assign(producto, updateProductoDto);
-   actualizarProducto.categoria = { id: updateProductoDto.idCategoria } as Categoria;
+    actualizarProducto.categoria = { id: updateProductoDto.idCategoria } as Categoria;
     return this.productosRepository.save(actualizarProducto);
   }
 
- async  remove(id: number) {
-  const producto = await this.findOne(id);
-  return this.productosRepository.delete(producto.id);
+  async remove(id: number) {
+    const producto = await this.findOne(id);
+    if (!producto) {
+      throw new NotFoundException(`La producto con el id ${id} no existe`);
+    }
+    return this.productosRepository.delete(producto.id);
   }
 }
